remove unnecessary unkey client
