+++ examples/nextjs-expiration/src/server/keys.ts
@@ -1,0 +1,7 @@
+type Key = {
+  key: string;
+  keyId: string;
+  expires: number;
+};
+
+export const keys: Key[] = [];
